requirement copy.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
contourpy==1.1.1
cvlib==0.2.7
cycler==0.12.1
cymem==2.0.8
fastai==2.7.17
fastcore==1.7.11
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.11.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
itsdangerous==2.2.0
jax==0.4.13
Jinja2==3.1.4
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
preshed==3.0.9
progressbar==2.5
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.0.4
wrapt==1.14.1
zipp==3.20.2
